refactor(web): split Posts browser shell
